[services, controllers] Include error handling that raise 403s whenever route requests miss necessary information, then incorporate that at the controller level

+++ services/predictions/destroyPredictions.js
@@ -4,9 +4,16 @@ const destroyRegressions = require('../regressions/destroyRegressions')
 
 async function destroyPredictions(source) {
     try {
-        await destroyRegressions(source)
-        await db.Prediction.deleteOne({source: source})
-        return 'Predictions deleted'
+        if (source) {
+            await destroyRegressions(source)
+            await db.Prediction.deleteOne({source: source})
+            return 'Predictions deleted'
+        } else {
+            throw {
+                code: 403,
+                message: 'Source must be provided'
+            }
+        }
     } catch (error) {
         throw error
     }
