+++ services/regressions/destroyRegressions.js
@@ -4,10 +4,17 @@ const regressionz = 'https://regressionz.herokuapp.com/api'
 
 async function destroyRegressions(source) {
     try {
-        await axios.delete(
-            regressionz + '?key=' + key + '&source=' + source
-        )
-        return 'Regressions deleted'
+        if (source) {
+            await axios.delete(
+                regressionz + '?key=' + key + '&source=' + source
+            )
+            return 'Regressions deleted'
+        } else {
+            throw {
+                code: 403,
+                message: 'Source must be provided'
+            }
+        }
     } catch (error) {
         throw error
     }
