+++ services/users/destroyUser.js
@@ -5,12 +5,19 @@ const readAllPredictions = require('../predictions/readAllPredictions')
 
 async function destroyUser(id) {
     try {
-        const allCollections = await readAllPredictions(id)
-        for (const collection of allCollections) {
-            await destroyPrediction(collection.prediction.source)
+        if (id) {
+            const allCollections = await readAllPredictions(id)
+            for (const collection of allCollections) {
+                await destroyPrediction(collection.prediction.source)
+            }
+            await db.User.deleteOne({_id: id})
+            return 'User deleted'
+        } else {
+            throw {
+                code: 403,
+                message: 'ID must be provided'
+            }
         }
-        await db.User.deleteOne({_id: id})
-        return 'User deleted'
     } catch (error) {
         throw error
     }
